MFC: r321689 Add a new "-N" option to umount(8), that does a forced dismount of an NFS mount point.
The new "-N" option does a forced dismount of an NFS mount point, but avoids
doing any checking of the mounted-on path, so that it will not get hung
when a vnode lock is held by another hung process on the mounted-on vnode.
The most common case of this is a "umount" with the "-f" option.
Other than avoiding checking the mounted-on path, it performs the same
forced dismount as a successful "umount -f" would do.

This commit includes a content change to the man page.


git-svn-id: https://svn.freebsd.org/base/stable/11@322910 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f