cam(4): Fix some warnings
When bcopy is treated as memcpy/memmove, Clang produces warnings that the
size argument doesn't match the type of the source.  This is true, it
doesn't match; we're aliasing the source.

Explicitly cast the source pointer to the expected type to remove the
warning.

No functional change.

Sponsored by:	Dell EMC Isilon


git-svn-id: https://svn.freebsd.org/base/head@323257 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f