MFC r321702,r321703:
r321702:

Load geom_gate(4) if necessary before running tests; skip if it can't be loaded

The test code prior to r311893 loaded geom_gate at test start if necessary and
skipped the tests if it couldn't be loaded.

The ATF-ifcation of this test done in r311893 unfortunately dropped this
functionality.

This change restores the geom_gate module load and skips the test(s) if unavailable
in an ATF-like way.

PR:		220164

r321703:

Remove superfluous `exit 0` added in r321702

atf_skip triggers equivalent functionality, which means the `exit 0`
is unreachable code.

PR:		220164
MFC with:	r321702


git-svn-id: https://svn.freebsd.org/base/stable/11@323300 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f